Fixed integer-truncation issue in 64-bit Clang builds
yajl_parse_integer should take length as a size_t not an unsigned int.
Fixes lloyd/yajl#107

+++ src/yajl_parser.c
@@ -33,7 +33,7 @@
 
  /* same semantics as strtol */
 long long
-yajl_parse_integer(const unsigned char *number, unsigned int length)
+yajl_parse_integer(const unsigned char *number, size_t length)
 {
     long long ret  = 0;
     long sign = 1;
